# const modulo optimization