--- a/dbms/src/Functions/modulo.cpp
+++ b/dbms/src/Functions/modulo.cpp
@@ -61,8 +61,20 @@ struct ModuloByConstantImpl
 
         /// Here we failed to make the SSE variant from libdivide give an advantage.
         size_t size = a.size();
-        for (size_t i = 0; i < size; ++i)
-            c[i] = a[i] - (a[i] / divider) * b; /// NOTE: perhaps, the division semantics with the remainder of negative numbers is not preserved.
+
+        /// strict aliasing optimization for char like arrays
+        auto * __restrict src = a.data();
+        auto * __restrict dst = c.data();
+        if (b & (b - 1))
+            for (size_t i = 0; i < size; ++i)
+                dst[i] = src[i] - (src[i] / divider) * b; /// NOTE: perhaps, the division semantics with the remainder of negative numbers is not preserved.
+        else
+        {
+            // gcc libdivide doesn't work well for pow2 division
+            auto mask = b - 1;
+            for (size_t i = 0; i < size; ++i)
+                dst[i] = src[i] & mask;
+        }
     }
 };