--- a/dbms/tests/performance/modulo.xml
+++ b/dbms/tests/performance/modulo.xml
@@ -1,0 +1,17 @@
+<test>
+    <type>loop</type>
+
+    <stop_conditions>
+        <any_of>
+            <iterations>10</iterations>
+        </any_of>
+    </stop_conditions>
+
+    <main_metric>
+        <min_time />
+    </main_metric>
+
+    <query>SELECT number % 128 FROM numbers(300000000) FORMAT Null</query>
+    <query>SELECT number % 255 FROM numbers(300000000) FORMAT Null</query>
+    <query>SELECT number % 256 FROM numbers(300000000) FORMAT Null</query>
+</test>