@akuzm 2019-11-18
https://github.com/ClickHouse/ClickHouse/commit/7084d090e805f76c660d14e1a7af162f6951fbd5

Style check was broken, fixed it.